requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
opencv-python==3.4.4.19
pandas==0.23.4
python-dateutil==2.7.5
python-nmap==0.6.1
pytz==2018.7
six==1.11.0
Werkzeug==0.14.1
